+++ packages/telestion-client-common/src/components/header/nav-bar/nav-bar.tsx
@@ -1,6 +1,6 @@
 import PropTypes from 'prop-types';
 import { Heading } from '@adobe/react-spectrum';
-import { Tabs, Item } from '@react-spectrum/tabs';
+import { Item, Tabs } from '@react-spectrum/tabs';
 import { useTitle } from '@wuespace/telestion-client-core';
 import { Link, useNavBarState } from './use-nav-bar-state';
 
@@ -81,7 +81,6 @@ export function NavBar({ links }: NavBarProps) {
 	// and hear on selection change instead
 	return (
 		<Tabs
-			height="100%"
 			items={items}
 			selectedKey={selected}
 			onSelectionChange={onSelectionChange}