+++ packages/telestion-client-prop-types/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.11.2](https://github.com/TelestionTeam/telestion-client/compare/v0.11.1...v0.11.2) (2021-03-09)
+
+**Note:** Version bump only for package @wuespace/telestion-client-prop-types
+
+
+
+
+
+### [0.11.1](https://github.com/TelestionTeam/telestion-client/compare/v0.11.0...v0.11.1) (2021-03-01)
+
+**Note:** Version bump only for package @wuespace/telestion-client-prop-types
+
+
+
+
+
+## [0.11.0](https://github.com/TelestionTeam/telestion-client/compare/v0.10.1...v0.11.0) (2021-02-28)
+
+**Note:** Version bump only for package @wuespace/telestion-client-prop-types
+
+
+
+
+
+### [0.10.1](https://github.com/TelestionTeam/telestion-client/compare/v0.10.0...v0.10.1) (2021-02-27)
+
+**Note:** Version bump only for package @wuespace/telestion-client-prop-types
+
+
+
+
+
 ## [0.10.0](https://github.com/TelestionTeam/telestion-client/compare/v0.9.0...v0.10.0) (2021-02-23)
 
 **Note:** Version bump only for package @wuespace/telestion-client-prop-types