+++ packages/telestion-client-types/CHANGELOG.md
@@ -3,6 +3,32 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [0.11.2](https://github.com/TelestionTeam/telestion-client/compare/v0.11.1...v0.11.2) (2021-03-09)
+
+
+### Bug Fixes
+
+* **types:** Allow undefined in generic props for widgets ([450bd38](https://github.com/TelestionTeam/telestion-client/commit/450bd38abb3255ebfe339f08f70a8f358b91f076))
+
+
+
+### [0.11.1](https://github.com/TelestionTeam/telestion-client/compare/v0.11.0...v0.11.1) (2021-03-01)
+
+
+### Bug Fixes
+
+* **types:** Add possible types to JSON serializable in object and array ([455a708](https://github.com/TelestionTeam/telestion-client/commit/455a708ae8118be9af9b8a79f4d5ec5500260bc9))
+
+
+
+## [0.11.0](https://github.com/TelestionTeam/telestion-client/compare/v0.10.1...v0.11.0) (2021-02-28)
+
+**Note:** Version bump only for package @wuespace/telestion-client-types
+
+
+
+
+
 ## [0.10.0](https://github.com/TelestionTeam/telestion-client/compare/v0.9.0...v0.10.0) (2021-02-23)
 
 **Note:** Version bump only for package @wuespace/telestion-client-types